Debugging and Error Checking
I fixed the following bugs:
Added if else statements for clicking, e.g. you haven't spun yet, etc.
Fixed a small logic error with winning, the script said if
(gameManager.gameIsDone) instead of the inverse.
Added the spin button to the prefab and it is disabled when the game is
over
Added a system which prevents you from landing on a completed face.